Upload edited runner that works on windows
runner_edit.py is an edited version of runner.py which is found in gym-locm\toolbox. both files run games of LOCM with two AI agents. my edit makes some changes to make it work on Windows and removes multiprocessing which was causing problems for me. There is also some code regarding printing actions (mostly commented out) but that can be ignored.